Change createEntityItem to float p.2

+++ CraftTweaker2-API/src/main/java/crafttweaker/api/item/IItemStack.java
@@ -447,7 +447,7 @@ public interface IItemStack extends IIngredient {
     boolean hasContainerItem();
     
     @ZenMethod
-    IEntityItem createEntityItem(IWorld world, int x, int y, int z);
+    IEntityItem createEntityItem(IWorld world, float x, float y, float z);
     
     @ZenMethod
     IEntityItem createEntityItem(IWorld world, IBlockPos pos);
